rationalize_extents: fix warnings

+++ src/flushthread.c
@@ -4399,7 +4399,7 @@ cont:
                 }
 
                 if (!(fcb->inode_item.flags & BTRFS_INODE_NODATASUM))
-                    add_checksum_entry(fcb->Vcb, er->address, er->skip_start / fcb->Vcb->superblock.sector_size, NULL, NULL);
+                    add_checksum_entry(fcb->Vcb, er->address, (ULONG)(er->skip_start / fcb->Vcb->superblock.sector_size), NULL, NULL);
 
                 ExAcquireResourceExclusiveLite(&er->chunk->lock, TRUE);
 
@@ -4453,7 +4453,7 @@ cont:
                 }
 
                 if (!(fcb->inode_item.flags & BTRFS_INODE_NODATASUM))
-                    add_checksum_entry(fcb->Vcb, er->address + er->length - er->skip_end, er->skip_end / fcb->Vcb->superblock.sector_size, NULL, NULL);
+                    add_checksum_entry(fcb->Vcb, er->address + er->length - er->skip_end, (ULONG)(er->skip_end / fcb->Vcb->superblock.sector_size), NULL, NULL);
 
                 ExAcquireResourceExclusiveLite(&er->chunk->lock, TRUE);
